+++ src/hotspot/cpu/ppc/continuationEntry_ppc.inline.hpp
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2022, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ *
+ */
+
+#ifndef CPU_PPC_CONTINUATIONENTRY_PPC_INLINE_HPP
+#define CPU_PPC_CONTINUATIONENTRY_PPC_INLINE_HPP
+
+#include "runtime/continuationEntry.hpp"
+
+// TODO: Implement
+
+inline frame ContinuationEntry::to_frame() const {
+  Unimplemented();
+  return frame();
+}
+
+inline intptr_t* ContinuationEntry::entry_fp() const {
+  Unimplemented();
+  return nullptr;
+}
+
+inline void ContinuationEntry::update_register_map(RegisterMap* map) const {
+  Unimplemented();
+}
+
+#endif // CPU_PPC_CONTINUATIONENTRY_PPC_INLINE_HPP
