+++ src/hotspot/share/runtime/thread.cpp
@@ -78,6 +78,8 @@
 #include "runtime/arguments.hpp"
 #include "runtime/atomic.hpp"
 #include "runtime/continuation.hpp"
+#include "runtime/continuationEntry.inline.hpp"
+#include "runtime/continuationHelper.inline.hpp"
 #include "runtime/fieldDescriptor.inline.hpp"
 #include "runtime/flags/jvmFlagLimit.hpp"
 #include "runtime/deoptimization.hpp"
@@ -2486,6 +2488,11 @@ javaVFrame* JavaThread::last_java_vframe(const frame f, RegisterMap *reg_map) {
     if (vf->is_java_frame()) return javaVFrame::cast(vf);
   }
   return NULL;
+}
+
+oop JavaThread::get_continuation() const {
+  assert(threadObj() != nullptr, "must be set");
+  return java_lang_Thread::continuation(threadObj());
 }
 
 Klass* JavaThread::security_get_caller_class(int depth) {
